requirements.txt
fastapi == 0.128.0
openai == 2.15.0
pyodbc == 5.3.0
pydantic == 2.12.5
uvicorn == 0.40.0
python-dotenv == 1.2.1